refactor: Coupon 구현체들 enum class 내부로 이동
